fix bugs. My sensors now cat start. All I2c sensora are working.
	modified:   flight/modules/Altitude/revolution/altitude.c
	modified:   flight/modules/FirmwareIAP/firmwareiap.c
	modified:   flight/pios/common/pios_bmp085_i2c.c
	modified:   flight/pios/common/pios_iap.c
	modified:   flight/pios/common/pios_l3g4200d.c
	modified:   flight/pios/common/pios_lps331ap.c
	modified:   flight/pios/inc/pios_bmp085_i2c.h
	modified:   flight/pios/inc/pios_lps331ap.h
	modified:   flight/targets/boards/brainyfly/board_hw_defs.c
	modified:   flight/targets/boards/brainyflybootloader/inc/pios_config.h
	modified:   flight/targets/boards/brainyfly/bootloader/main.c
	modified:   flight/targets/boards/brainyfly/firmware/inc/pios_config.h
	modified:   flight/targets/boards/brainyfly/firmware/pios_board.c

+++ flight/pios/inc/pios_bmp085_i2c.h
@@ -76,7 +76,7 @@ extern void PIOS_BMP085_Init(uint32_t i2c_id);
 extern void PIOS_BMP085_StartADC(ConversionTypeTypeDef Type);
 extern void PIOS_BMP085_ReadADC(void);
 extern int16_t PIOS_BMP085_GetTemperature(void);
-extern int32_t PIOS_BMP085_GetPressure(void);
+extern float PIOS_BMP085_GetPressure(void); // kpa
 extern bool PIOS_BMP085_Read(uint8_t address, uint8_t *buffer, uint8_t len);
 extern bool PIOS_BMP085_Write(uint8_t address, uint8_t buffer);
 extern int32_t PIOS_BMP085_Test();
